未测试

+++ src/main/java/com/Dao/UserRepository.java
@@ -119,8 +119,8 @@ public class UserRepository {
     //注销(个人)
     public boolean deleteUser(int userId) {
         try {
-            //删除二级评论
-            template.update("delete from Comment where ");
+            //删除回复该用户的评论
+            template.update("delete from Comment where commentId in (select commentOwnerId from User,Comment where commentReplyName=userName and userId=?)", userId);
             //删除用户的评论
             template.update("delete from Comment where commentOwnerId=?", userId);
             //删除用户创建的帖子
